+++ vck-openid/src/commonMain/kotlin/at/asitplus/wallet/lib/openid/OpenId4VpRequestOptions.kt
@@ -1,5 +1,6 @@
 package at.asitplus.wallet.lib.openid
 
+import at.asitplus.data.validation.third_party.kotlin.collections.requireIsNotNullOrEmpty
 import at.asitplus.openid.OpenIdConstants
 import at.asitplus.openid.OpenIdConstants.ResponseMode
 import at.asitplus.openid.OpenIdConstants.SCOPE_OPENID
@@ -101,7 +102,10 @@ data class OpenId4VpRequestOptions(
         if (isAnyDcApi) {
             require(isDcql) { "DC API only supports DCQL" }
             require(!isSiop) { "DC API does not support SIOP (id_token)" }
-            requireNotNull(expectedOrigins) { "Expected origins must be set for DC API" }
+            if (populateClientId) {
+                // should be a signed DC API request if client_id has to be assigned
+                expectedOrigins.requireIsNotNullOrEmpty { "Expected origins must be set for DC API" }
+            }
         } else {
             require(populateClientId) { "client_id should be set for anything but (unsigned) DC API requests" }
         }